+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Renamed internal `main` symbol to `__main` to avoid clashes with symbols from the application.
 
+### Improvements
+
+- Added secondary core entry point.
+
 ## 0.1.1
 
 ### Improvements